requirements.txt
# --- Web Server ---
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1

# --- Machine Learning---
tensorflow==2.16.1
scikit-learn==1.6.1
pandas==2.2.0
numpy==1.26.4
joblib==1.3.2
huggingface_hub==0.21.3

# --- Scraper & DB ---
requests==2.31.0
beautifulsoup4==4.12.3
psycopg2-binary==2.9.9